requirements.txt
requests
pymongo
python-dotenv
pandas
git+https://github.com/rongardF/tvdatafeed.git